recorder/requirements.txt
Flask==3.0.3
pywebview==4.4
